pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "google-api-python-client>=2.163.0",
    "oauth2client>=4.1.3",
    "schedule>=1.2.2",
    "pillow>=11.1.0",
    "werkzeug>=3.1.3",
    "instagrapi>=2.1.3",
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.8.4",
    "sqlalchemy>=2.0.38",
    "wtforms>=3.2.1",
]
